+++ packages/eyes-api/src/Eyes.ts
@@ -1,8 +1,8 @@
 import type * as types from '@applitools/types'
 import * as utils from '@applitools/utils'
 import {SessionType, SessionTypeEnum} from './enums/SessionType'
-import {StitchModeEnum} from './enums/StitchMode'
-import {MatchLevelEnum} from './enums/MatchLevel'
+import {StitchMode, StitchModeEnum} from './enums/StitchMode'
+import {MatchLevel, MatchLevelEnum} from './enums/MatchLevel'
 import {EyesError} from './errors/EyesError'
 import {NewTestError} from './errors/NewTestError'
 import {DiffsFoundError} from './errors/DiffsFoundError'
@@ -562,7 +562,7 @@ export class Eyes<TDriver = unknown, TElement = unknown, TSelector = unknown> {
   getMatchLevel(): MatchLevelEnum {
     return this._config.getMatchLevel()
   }
-  setMatchLevel(matchLevel: MatchLevelEnum) {
+  setMatchLevel(matchLevel: MatchLevel) {
     this._config.setMatchLevel(matchLevel)
   }
 
@@ -655,7 +655,7 @@ export class Eyes<TDriver = unknown, TElement = unknown, TSelector = unknown> {
   getStitchMode(): StitchModeEnum {
     return this._config.getStitchMode()
   }
-  setStitchMode(stitchMode: StitchModeEnum) {
+  setStitchMode(stitchMode: StitchMode) {
     this._config.setStitchMode(stitchMode)
   }