+++ packages/eyes-api/src/input/Configuration.ts
@@ -1,10 +1,10 @@
 import * as utils from '@applitools/utils'
 import {SessionType, SessionTypeEnum} from '../enums/SessionType'
 import {StitchMode, StitchModeEnum} from '../enums/StitchMode'
-import {MatchLevelEnum} from '../enums/MatchLevel'
-import {BrowserTypeEnum} from '../enums/BrowserType'
-import {DeviceNameEnum} from '../enums/DeviceName'
-import {ScreenOrientationEnum} from '../enums/ScreenOrientation'
+import {MatchLevel, MatchLevelEnum} from '../enums/MatchLevel'
+import {BrowserType, BrowserTypeEnum} from '../enums/BrowserType'
+import {DeviceName} from '../enums/DeviceName'
+import {ScreenOrientation, ScreenOrientationEnum} from '../enums/ScreenOrientation'
 import {AccessibilitySettings} from './AccessibilitySettings'
 import {DesktopBrowserInfo, ChromeEmulationInfo, IOSDeviceInfo, ChromeEmulationInfoLegacy} from './RenderInfo'
 import {CutProvider} from './CutProvider'
@@ -270,7 +270,7 @@ export class ConfigurationData<TElement = unknown, TSelector = unknown>
   getSessionType(): SessionTypeEnum {
     return this.sessionType as SessionTypeEnum
   }
-  setSessionType(sessionType: SessionTypeEnum): this {
+  setSessionType(sessionType: SessionType): this {
     this.sessionType = sessionType
     return this
   }
@@ -719,7 +719,7 @@ export class ConfigurationData<TElement = unknown, TSelector = unknown>
   getMatchLevel(): MatchLevelEnum {
     return this.defaultMatchSettings?.matchLevel as MatchLevelEnum
   }
-  setMatchLevel(matchLevel: MatchLevelEnum): this {
+  setMatchLevel(matchLevel: MatchLevel): this {
     if (!this.defaultMatchSettings) this.defaultMatchSettings = {}
     this.defaultMatchSettings.matchLevel = matchLevel
     return this
@@ -804,7 +804,7 @@ export class ConfigurationData<TElement = unknown, TSelector = unknown>
   getStitchMode(): StitchModeEnum {
     return this.stitchMode as StitchModeEnum
   }
-  setStitchMode(stitchMode: StitchModeEnum): this {
+  setStitchMode(stitchMode: StitchMode): this {
     this.stitchMode = stitchMode
     return this
   }
@@ -961,12 +961,12 @@ export class ConfigurationData<TElement = unknown, TSelector = unknown>
     return this
   }
   addBrowser(browserInfo: RenderInfo): this
-  addBrowser(width: number, height: number, name?: BrowserTypeEnum): this
-  addBrowser(browserInfoOrWidth: RenderInfo | number, height?: number, name = BrowserTypeEnum.CHROME) {
+  addBrowser(width: number, height: number, name?: BrowserType): this
+  addBrowser(browserInfoOrWidth: RenderInfo | number, height?: number, name: BrowserType = BrowserTypeEnum.CHROME) {
     if (utils.types.isObject(browserInfoOrWidth)) return this.addBrowsers(browserInfoOrWidth)
     else return this.addBrowsers({width: browserInfoOrWidth, height, name})
   }
-  addDeviceEmulation(deviceName: DeviceNameEnum, screenOrientation = ScreenOrientationEnum.PORTRAIT) {
+  addDeviceEmulation(deviceName: DeviceName, screenOrientation: ScreenOrientation = ScreenOrientationEnum.PORTRAIT) {
     if (!this.browsersInfo) this.browsersInfo = []
     this.browsersInfo.push({chromeEmulationInfo: {deviceName, screenOrientation}})
     return this