+++ packages/eyes-api/CHANGELOG.md
@@ -1,8 +1,9 @@
 # Changelog
 
 ## Unreleased
-- updated to @applitools/types@1.0.9 (from 1.0.7)
 
+- allow string literals to be set in data classes and fluent api
+- updated to @applitools/types@1.0.9 (from 1.0.7)
 
 ## 1.0.12 - 2021/8/13