+++ packages/eyes-api/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+
+## 1.0.13 - 2021/8/31
+
 - allow string literals to be set in data classes and fluent api
 - updated to @applitools/types@1.0.9 (from 1.0.7)
 - updated to @applitools/types@1.0.10 (from 1.0.9)